Scripts/requirements.txt
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dnspython==1.16.0
flake8==4.0.1
idna==3.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.19.5
pandas==1.4.0
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
pymongo==3.11.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
schedule==1.1.0
six==1.16.0
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
